requirements.txt
asgiref==3.6.0
Authlib==1.2.0
boto3==1.26.71
botocore==1.29.71
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.0
defusedxml==0.7.1
Django==4.1.5
django-phonenumber-field==7.0.2
django-recaptcha==3.0.0
django-storages==1.13.2
djangorestframework==3.14.0
ecdsa==0.18.0
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
oauthlib==3.2.2
phonenumbers==8.13.7
Pillow==9.4.0
psycopg2-binary==2.9.5
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.1
python-http-client==3.3.7
python-jose==3.3.0
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
sendgrid==6.9.7
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
tweepy==4.12.1
twilio==7.16.4
urllib3==1.26.14
whitenoise==6.3.0
